api/requirements.txt
asgiref==3.2.5
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
mysqlclient==1.4.6
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
